backend/requirements.txt
# =============================================================================
# CORE FRAMEWORK DEPENDENCIES
# =============================================================================

# FastAPI - Modern, fast web framework for building APIs
fastapi==0.115.6

# Uvicorn - ASGI server for FastAPI with standard extras
uvicorn[standard]==0.32.1

# =============================================================================
# SPOTIFY INTEGRATION
# =============================================================================

# Spotipy - Python library for Spotify Web API
spotipy==2.24.0

# =============================================================================
# LLM INTEGRATION (OpenAI)
# =============================================================================

# OpenAI - OpenAI API client for GPT models
openai>=1.0.0

# Hugging Face - Free AI models
huggingface_hub>=0.16.0

# Google Gemini - Free AI model
google-generativeai>=0.3.0

# LiteLLM - Multi-provider LLM integration
litellm>=1.0.0

# Redis - For caching user profiles (optional for future use)
redis>=4.0.0

# =============================================================================
# ENVIRONMENT & CONFIGURATION
# =============================================================================

# Python-dotenv - Load environment variables from .env files
python-dotenv==1.0.1

# Itsdangerous - Secure data serialization for sessions
itsdangerous==2.2.0

# =============================================================================
# DATA VALIDATION & PROCESSING
# =============================================================================

# Pydantic - Data validation using Python type annotations
pydantic==2.10.4

# Requests - HTTP library for API calls
requests==2.32.3
